\name{data}
\alias{data}
\docType{data}
\title{
  A simulated dataset.
}
\description{
  The dataset contains the information on stratum, disease status, SNP genotype, and three covariates.
}
\usage{data(data)}
\format{
  A data frame containing the following variables for each of 1200 observations: status (disease status coded as 1 for case or 0 for control), stratum (a three-level stratum variable coded as 1, 2, or 2), covariate.1, covariate.2, covariate.3 (three continuous covariates), SNP.1,...,SNP.100 (100 SNP genotypes coded as 0, 1, or 2 according to the number of minor alleles).
}
\examples{
  data(data);
}
\keyword{datasets}
