\name{data}
\alias{data}
\docType{data}
\title{
  A simulated dataset.
}
\description{
  The dataset contains the information on stratum, disease status, SNP genotype, and three covariates.
}
\usage{data(data)}
\format{
  A data frame containing the following information for each of 1200 observations:
  \describe{
    \item{\code{stratum}}{a numeric vector}
    \item{\code{status}}{a numeric vector}
    \item{\code{genotype}}{a numeric vector}
    \item{\code{covariate.1}}{a numeric vector}
    \item{\code{covariate.2}}{a numeric vector}
    \item{\code{covariate.3}}{a numeric vector}
  }
}
\examples{
  data(data);
}
\keyword{datasets}
